// src/components/StoryView/types.ts
export enum StroyTypes {
  Image = 'image',
  Video = 'video',
  Text = 'text',
}
// ...
export interface StoryViewProps {
setStories:any
  storyDataLoaded?:boolean
  pause?: boolean;
  onVideoLoaded?: (arg0: OnLoadData) => void;
  videoDuration: number[];
  onVideoProgress?: (progressData?: OnProgressData) => void;
  onVideoEnd?: () => void;
  onImageLoaded?: () => void;
  stories: StoryType[];
  showSourceIndicator?: boolean;
  sourceIndicatorProps?: ActivityIndicatorProps;
  videoProps?: VideoProperties;
  index?: number;
  storyIndex?: number;
  setComments?:any
  viewStoryFuction?:any
  textStyle?:any
}
// ...
export type StoryType = {
story: any;
  text_story_properties: any;
  resolved: boolean;
  id: number;
  url?: string;
  thumbnail?: string;
  type?: string | 'image' | 'video' | 'text';
  duration?: number;
  isReadMore?: boolean;
  isSeen?: boolean;
  isPaused?: boolean;
  created?: string;
  storyId?: number;
  title?: string;
};